+++ src/engine/alt/nbio.c
@@ -865,26 +865,42 @@ static int opennonblock_tgt(lua_State* L, bool wr)
 	if (vobj->feed.state.tag != ARCAN_TAG_FRAMESERV)
 		arcan_fatal("open_nonblock(tgt), target must be a valid frameserver.");
 
-	int outp[2];
-	if (-1 == pipe(outp)){
-		arcan_warning("open_nonblock(tgt), pipe-pair creation failed: %d\n", errno);
-		return 0;
-	}
-
+/* overloaded form:
+ *  open_nonblock(vid, r | w, type, nbio_ud)
+ *
+ *  This takes an existing userdata, extracts the descriptor and sends to the
+ *  target, while disassociating the descriptor from the argument source.
+ */
 	const char* type = luaL_optstring(L, 3, "stream");
-
-/* WRITE mode = 'INPUT' in the client space */
-	int dst = wr ? outp[0] : outp[1];
-	int src = wr ? outp[1] : outp[0];
-
-/* in any scenario where this would fail, "blocking" behavior is acceptable */
-	alt_nbio_nonblock_cloexec(src, true);
 	struct arcan_event ev = {
 		.category = EVENT_TARGET,
 		.tgt.kind = wr ? TARGET_COMMAND_BCHUNK_IN : TARGET_COMMAND_BCHUNK_OUT
 	};
 	snprintf(ev.tgt.message, COUNT_OF(ev.tgt.message), "%s", type);
+	if (lua_type(L, 4) == LUA_TUSERDATA){
+		struct nonblock_io** ibb = luaL_checkudata(L, 4, "nonblockIO");
+		struct nonblock_io* ib = *ibb;
 
+		if (ib->fd > 0){
+			platform_fsrv_pushfd(fsrv, &ev, ib->fd);
+			close(ib->fd);
+			ib->fd = -1;
+		}
+
+		return 0;
+	}
+
+/* WRITE mode = 'INPUT' in the client space */
+	int outp[2];
+	if (-1 == pipe(outp)){
+		arcan_warning("open_nonblock(tgt), pipe-pair creation failed: %d\n", errno);
+		return 0;
+	}
+	int dst = wr ? outp[0] : outp[1];
+	int src = wr ? outp[1] : outp[0];
+
+/* in any scenario where this would fail, "blocking" behavior is acceptable */
+	alt_nbio_nonblock_cloexec(src, true);
 	if (ARCAN_OK != platform_fsrv_pushfd(fsrv, &ev, dst)){
 		close(dst);
 		close(src);