+++ src/platform/stub/event.c
@@ -28,7 +28,7 @@ void platform_event_analogall(bool enable, bool mouse)
 {
 }
 
-bool platform_event_translation(
+int platform_event_translation(
 	int devid, int action, const char** names, const char** err)
 {
 	*err = "Unsupported";