+++ src/platform/wscons/event.c
@@ -319,7 +319,7 @@ void platform_event_keyrepeat(arcan_evctx* ctx, int* period, int* delay)
 	}
 }
 
-bool platform_event_translation(
+int platform_event_translation(
 	int devid, int action, const char** names, const char** err)
 {
 	*err = "Unsupported";